修复批量删除和拖动排序的问题

+++ src/main/java/com/hashtech/entity/ResourceTableEntity.java
@@ -74,7 +74,7 @@ public class ResourceTableEntity implements Serializable {
     /**
      * 表数据量
      */
-    private Integer dataSize;
+    private Long dataSize;
 
     /**
      * 注意事项
